Post-refactor fixes
