+++ q2_moshpit/eggnog/dbs.py
@@ -7,28 +7,30 @@
 # ----------------------------------------------------------------------------
 import os
 import shutil
+
 import pandas as pd
 from qiime2.core.exceptions import ValidationError
-from q2_types.feature_data import ProteinSequencesDirectoryFormat
-from q2_types.reference_db import (
-    EggnogRefDirFmt, DiamondDatabaseDirFmt, NCBITaxonomyDirFmt,
-    EggnogProteinSequencesDirFmt
-)
-from q2_types.profile_hmms import (
-    ProteinMultipleProfileHmmDirectoryFmt,
-    PressedProfileHmmsDirectoryFmt
-)
-from q2_types.genome_data import ProteinsDirectoryFormat
+
 from q2_moshpit._utils import (
     run_command, _process_common_input_params, colorify,
     _calculate_md5_from_file
 )
-from q2_moshpit.eggnog._utils import (
+from q2_moshpit.eggnog.types import EggnogHmmerIdmapDirectoryFmt
+from q2_moshpit.eggnog.utils import (
     _parse_build_diamond_db_params, _download_and_build_hmm_db,
     _download_fastas_into_hmmer_db,
     _validate_eggnog_hmmer_taxon_id
 )
-from q2_moshpit.eggnog._format import EggnogHmmerIdmapDirectoryFmt
+from q2_types.feature_data import ProteinSequencesDirectoryFormat
+from q2_types.genome_data import ProteinsDirectoryFormat
+from q2_types.profile_hmms import (
+    ProteinMultipleProfileHmmDirectoryFmt,
+    PressedProfileHmmsDirectoryFmt
+)
+from q2_types.reference_db import (
+    EggnogRefDirFmt, DiamondDatabaseDirFmt, NCBITaxonomyDirFmt,
+    EggnogProteinSequencesDirFmt
+)
 
 
 def fetch_eggnog_db() -> EggnogRefDirFmt:
