+++ q2_moshpit/eggnog/types/tests/test_types.py
@@ -6,7 +6,9 @@
 # The full license is in the file LICENSE, distributed with this software.
 # ----------------------------------------------------------------------------
 from qiime2.plugin.testing import TestPluginBase
-from q2_moshpit.eggnog import EggnogHmmerIdmap, EggnogHmmerIdmapDirectoryFmt
+from q2_moshpit.eggnog.types import (
+    EggnogHmmerIdmap, EggnogHmmerIdmapDirectoryFmt
+)
 
 
 class TestEggnogHmmerIdmap(TestPluginBase):
